pyproject.toml
[tool.poetry]
name = "goodreads"
version = "0.1.0"
description = ""
authors = ["Yaser Khahani <yaserkh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.0.7,<5.1"
djangorestframework = ">=3.15.2,<3.16"
psycopg2-binary = ">=2.9.9, <2.10"
djangorestframework-simplejwt = ">=5.3.1,<5.4"
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
